Update findAndMount.sh for RONIN LINK
Some recommendations for RONIN Link integration: Want the user to be able to automatically mount partitioned drives that don't require filesystem extension (i.e. restored from backup but not resized) but identify partitioned drives that do require a filesystem extension (i.e. drive resized when restored from a backup - user will need to perform extension themselves but we can tell them which drives need this) - I have added this identification to the script so they are stored in two separate variables now. Have also added some RONIN LINK lines that will assist with displaying the information for each of the drives so that we can read it in and create the required html elements dynamically - IDs will match up with RONIN interface e.g. /dev/sdd rather than nvme1n1 etc. Partition and mount drive function will likely need to be split into two functions and put into separate scripts for RONIN LINK - mount only script (for drives that already have a file system or full partition and simply need to be mounted) and partition and mount script (for empty EBS Drives and ephemeral drives that require both). Not sure whether we can add the functionality to determine whether a drive is General, Magnetic, Hot, Cold for RONIN LINK - from what I've read we don't have the required permissions on the machines to do this?